+++ sources/eth_benchmark.py
@@ -1,0 +1,12 @@
+import speedtest
+
+def eth_benchmark():
+    global download
+    global upload
+    global ping
+
+    st = speedtest.Speedtest()
+    st.get_best_server()
+    download = round(st.download()/1e6, 1)  # Unit: Mbps
+    upload = round(st.upload()/1e6, 1)  # Unit: Mbps
+    ping = round(st.results.ping, 1)  # Unit: ms
